+++ sig.cpp
@@ -554,7 +554,7 @@ int dline_add(dline_t * dl, ea_t ea, char options)
 
 	if (options)
 	{
-		qsnprintf(addr, sizeof(addr), "%u", (unsigned int)ea);
+		qsnprintf(addr, sizeof(addr), "%lu", (unsigned long)ea);
 		out_snprintf("%s ", addr);
 	}
 
@@ -812,7 +812,7 @@ psig_t * sig_class_generate(ea_t ea)
 	sig_set_start(sig, ea);
 
 	// Adds function name
-	qsnprintf(buf, sizeof(buf), "sub_%u", (unsigned int)ea);
+	qsnprintf(buf, sizeof(buf), "sub_%lu", (unsigned long)ea);
 	sig_set_name(sig, buf);
 
 	// Adds class references
